Some people use a hacked libtool that doesn't even install .la files

+++ sanitize-la.sh
@@ -1,3 +1,6 @@
 #!/bin/sh
 
-sed "s/dependency_libs=.*/dependency_libs=''/" < $1 > $1T && mv $1T $1
+if [ -f $1 ]; then
+    sed "s/dependency_libs=.*/dependency_libs=''/" < $1 > $1T && mv $1T $1
+fi
+
